Refine CLI runtime structure
Split the CLI library into focused command and configuration modules, simplify the execution flow to lower complexity, and extend the tests and docs per review feedback.
