+++ Cargo.toml
@@ -20,6 +20,7 @@ rstest = "0.23"
 rstest-bdd = "0.1.0-alpha4"
 rstest-bdd-macros = "0.1.0-alpha4"
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 thiserror = "2.0"
 tempfile = "3.10"
 dirs = "6.0"
