+++ crates/weaver-cli/Cargo.toml
@@ -7,7 +7,7 @@ version.workspace = true
 weaver-config = { path = "../weaver-config", features = ["cli"] }
 clap = { workspace = true }
 serde = { workspace = true }
-serde_json = "1.0"
+serde_json = { workspace = true }
 ortho_config = { workspace = true }
 thiserror = { workspace = true }
 socket2 = { version = "0.5", features = ["all"] }
