+++ docs/users-guide.md
@@ -111,7 +111,7 @@ does not accept a request within that window, the CLI aborts with a descriptive
 error instead of hanging. Likewise, if the daemon sends ten consecutive blank
 lines, the CLI emits a warning, stops reading further, and reports failure
 unless an exit status was already observed. Any session that ends without an
-explicit exit message is treated as an error so callers do not misinterpret a
+explicit exit message is treated as an error, so callers do not misinterpret a
 partial response as success.
 
 ### Capability probe
